pyproject.toml
[project]
name = "translator-server"
version = "0.1.0"
description = "FastAPI translation service powered by fastText language detection and local M2M100 translation."
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "fasttext-wheel>=0.9.2",
    "numpy<2",
    "sentencepiece>=0.2.0",
    "setuptools<81",
    "torch>=2.4.0",
    "transformers>=4.46.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "zhconv>=1.4.3",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
]

[project.scripts]
translator-server = "translator_server.main:run"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/translator_server"]

[tool.uv]
package = true
